deprecated 4 narrower concepts to normalise vocab
This vocab contains many top concepts but hardly any narrower concepts.  Ideal state is for more granularity across all of the survey types but this is a large body of work that has not been completed. Until then it should stay relatively flat. Concepts just #'d out for easy reinstatement in the future.