added BeforePresentCalibrated timescale and other minor edits
The expand for Before Present did not include "Before Present" so I added a skos:broader line to that entry. I hope that is correct. I think that the expand should also be labeled "Geologic and Archeologic Timescales" or something like that. I have never done a pull request before, so I hope this one comes through.