Update geologic-feature-types.ttl
I hope that these changes suffice.
Note:
1. I have not inserted skos notations
2. The definition of an igneous subprovince is outstanding (i.e. none provided, not otherwise)
3. Ridge was included previously (see definition), but I am thinking that ridge refers to a basement ridge and should be modified accordingly. Advice please.
4. I have not included:
•	OBSOLETE (MERLIN)
Obsolete
Structural unit terms that are considered to be obsolete and have been superceded by other terms.
Scope:Note: It has never been possible to delete records/terms from authority tables, and flagging a term as being obsolete is necessary. Care should be taken in deleting a record because it may have been linked to in Surface Geology.

•	MISC (MERLIN)
Not to be used. Do not include in ttl.
Terms in MERLIN that fell here (only three) have been now flagged as MISC (by Ian).

•	NOTQLD (MERLIN)
Forget for the moment: Not in Qld.
Comment from Ian: What is not obvious is why they are in the table at all. There are 11 of these, and include things like Otway basin, Bonaparte basin, Canning Basin etc. All are petroleum basins and possibly were in the codes that came across to CPF_STRUCTURAL_UNITS from the old QPED database that was migrated to BOREHOLE. Check if they are needed in BOREHOLE and if not delete them and the code.

 Should I do so??