Update earth-science-data-category.ttl
Today's additions and edits to the ESDC vovab. I have kept it simple and will continue to do so.

Let me know if you're happy with the additions, and I will add them to the lodgement list. Ran out of time to do so today, except for getting rid of the diphthong in the spelling of paleontology (and in the vocab itself). 
@predam1 @SalEdwards01 for your info, and what suggestions/requests do you have, but keeping it simple please.