Update registryroles.ttl
@KellyVance @katewd 
Not sure what's going on here. Definitions are long; I have contracted them slightly. But, they may need more work, including partial transfer to a scopeNote?
In the coding, it looks like '\n' is supposed to form a new line?? But, I have deleted it!! Advice here please.
Also, I have left untouched '&#39' (in the second definition that I have edited). What on Earth does it mean or what is it supposed to represent??