Update geologic-feature-types.ttl
I have edited this branch, largely to convert dcterms:source to dcterms:references, to accord with your critique Nick, but I still cannot get your validator to work for me.
There are other edits in here too. I have fixed up some of the referencing, but much more work is needed on this to get the formatting consistent within the vocab (and across vocabs). This work will sit on the backburner for the present.