Update gsq-feature-status.ttl
@BlakePaul @predam1 
Hi Paul, Micaela,
I have edited the branch for this PR that was sent to you a week or more ago, as there was an error in line 25, and I have updated the info email address therein. Paul, in terms of your requirements for RU status (in your email from a long time ago), I restricted the list for status to what you see in the file. Would you both kindly advise on this please, to see what needs to be edited, added or, in terms of what is/was in MERLIN, disregarded (or otherwise merged with a term in the said list).
Cheers,
John