Update geologic-feature-types.ttl
@nicholascar @predam1 @mckillopm @LizDerrington @CourteneyD @vladrizm @Steven-Bowden 

I Have added the upper level framework for concepts under georesource accumulation ... and, Andrew, I have also added the terms that you required for petroleum. Placement of resource accumulation, however, remains open to discussion, whether in geologic-feature types or separate from this ontology. This work was undertaken in last November, before I went on leave.

Potentially, it is up to Derek, Vlad & Courtney, to develop what is needed for the ontology for Minerals; Micaela & Mike, for coal; and Andrew & Liz, for petroleum. Suggestions are welcome here, as above for placement. The separate ontologies for each resource type need to have a common theme and a purpose that meets our/your needs.